Move AttributeAssignmentMixin and validate_attributes_input to attribute module
